refactor: extract strategy and account side panels from main.tsx (Block E)
