+++ web/console/src/main.tsx
@@ -28,6 +28,8 @@ import { OrdersPanel } from './panels/OrdersPanel';
 import { AlertsPanel } from './panels/AlertsPanel';
 import { PositionsPanel } from './panels/PositionsPanel';
 import { FillsPanel } from './panels/FillsPanel';
+import { StrategySidePanel } from './pages/StrategySidePanel';
+import { AccountSidePanel } from './pages/AccountSidePanel';
 
 import { AccountSummary, AccountRecord, StrategyVersion, StrategyRecord, AccountEquitySnapshot, Order, Fill, Position, PaperSession, LiveSession, ChartCandle, ChartAnnotation, MarkerLegendItem, BacktestRun, BacktestOptions, LiveAdapter, SignalSourceDefinition, SignalSourceCatalog, SignalSourceType, SignalBinding, SignalRuntimeAdapter, SignalRuntimeSession, ReplayReasonStats, ReplaySample, ExecutionTrade, SourceFilter, EventFilter, TimeWindow, MarkerDetail, ChartOverrideRange, SelectedSample, SelectableSample, RuntimeMarketSnapshot, RuntimeSourceSummary, RuntimeReadiness, SignalBarCandle, AlertItem, PlatformAlert, PlatformNotification, TelegramConfig, RuntimePolicy, LivePreflightSummary, LiveNextAction, LiveDispatchPreview, LiveSessionExecutionSummary, LiveSessionHealth, HighlightedLiveSession, LiveSessionFlowStep, SessionMarker, AuthSession } from './types/domain';
 import { sampleStatus, buildLinePath, summarizeRange, summarizeTimeRange, filterChartAnnotations, matchesEventFilter, resolveChartAnchor, buildTimeRange, buildSampleRange, buildSampleKey, annotationMatchesSample, findNearestAnnotation, toMarkerDetail, markerShape, markerPosition, markerColor, markerText, annotationTone, paperAccountsFromSummaries, strategyLabel, getNumber, getRecord, getList, deriveRuntimeMarketSnapshot, deriveRuntimeSourceSummary, deriveRuntimeReadiness, deriveSignalBarCandles, mapChartCandlesToSignalBarCandles, applyDefaultChartWindow, derivePrimarySignalBarState, buildRuntimeEventNotes, buildSourceStateNotes, buildSignalBarDecisionNotes, buildSignalBarStateNotes, deriveSignalActionSummary, deriveLivePreflightSummary, deriveLiveDispatchPreview, deriveLiveSessionExecutionSummary, derivePaperSessionExecutionSummary, deriveSessionMarkers, deriveLiveSessionHealth, deriveHighlightedLiveSession, deriveLiveSessionFlow, liveSessionHealthPriority, deriveLiveNextAction, liveSessionHealthTone, buildSignalActionNotes, buildTimelineNotes, summarizeOrderPreflight, derivePaperAlerts, deriveLiveAlerts, dedupeAlerts, buildAlertNotes, alertLevelTone, alertScopeTone, telegramDeliveryTone, runtimeReadinessTone, decisionStateTone, signalKindTone, signalActionTone, boolTone, boolLabel } from './utils/derivation';
@@ -1809,870 +1811,17 @@ function App() {
           </span>
         </div>
       }
-            sidePanelContent={
+      sidePanelContent={
         sidebarTab === 'strategy' ? (
-          <div className="p-4 space-y-6">
-            <section id="backtests" className="panel panel-backtests">
-          <div className="panel-header">
-            <div>
-              <p className="panel-kicker">Backtests</p>
-              <h3>回测配置与运行记录</h3>
-            </div>
-            <div className="range-box">
-              <span>{backtests.length} runs</span>
-              <span>{strategies.length} strategies</span>
-            </div>
-          </div>
-          <div className="backtest-layout">
-            <div className="backtest-form">
-              <div className="form-grid">
-                <label className="form-field">
-                  <span>Strategy Version</span>
-                  <select
-                    value={backtestForm.strategyVersionId}
-                    onChange={(event) => setBacktestForm((current: any) => ({ ...current, strategyVersionId: event.target.value }))}
-                  >
-                    {strategies.map((strategy) => (
-                      <option key={strategy.id} value={strategy.currentVersion?.id ?? ""}>
-                        {strategyLabel(strategy)}
-                      </option>
-                    ))}
-                  </select>
-                </label>
-                <label className="form-field">
-                  <span>Signal Timeframe</span>
-                  <select
-                    value={backtestForm.signalTimeframe}
-                    onChange={(event) => setBacktestForm((current: any) => ({ ...current, signalTimeframe: event.target.value }))}
-                  >
-                    {(backtestOptions?.signalTimeframes ?? ["4h", "1d"]).map((item) => (
-                      <option key={item} value={item}>
-                        {item}
-                      </option>
-                    ))}
-                  </select>
-                </label>
-                <label className="form-field">
-                  <span>Execution Source</span>
-                  <select
-                    value={backtestForm.executionDataSource}
-                    onChange={(event) => setBacktestForm((current: any) => ({ ...current, executionDataSource: event.target.value }))}
-                  >
-                    {(backtestOptions?.executionDataSources ?? ["tick", "1min"]).map((item) => (
-                      <option key={item} value={item}>
-                        {item}
-                      </option>
-                    ))}
-                  </select>
-                </label>
-                <label className="form-field">
-                  <span>Symbol</span>
-                  <input
-                    value={backtestForm.symbol}
-                    onChange={(event) => setBacktestForm((current: any) => ({ ...current, symbol: event.target.value.toUpperCase() }))}
-                    placeholder="BTCUSDT"
-                  />
-                </label>
-                <label className="form-field">
-                  <span>From (RFC3339)</span>
-                  <input
-                    value={backtestForm.from}
-                    onChange={(event) => setBacktestForm((current: any) => ({ ...current, from: event.target.value }))}
-                    placeholder="2020-01-01T00:00:00Z"
-                  />
-                </label>
-                <label className="form-field">
-                  <span>To (RFC3339)</span>
-                  <input
-                    value={backtestForm.to}
-                    onChange={(event) => setBacktestForm((current: any) => ({ ...current, to: event.target.value }))}
-                    placeholder="2020-01-31T23:59:59Z"
-                  />
-                </label>
-              </div>
-              <div className="backtest-actions">
-                <ActionButton
-                  label={backtestAction ? "Submitting..." : "Create Backtest"}
-                  disabled={
-                    backtestAction ||
-                    backtestForm.strategyVersionId.trim() === "" ||
-                    backtestForm.symbol.trim() === "" ||
-                    selectedExecutionAvailability === "missing" ||
-                    !selectedSymbolAvailable
-                  }
-                  onClick={createBacktestRun}
-                />
-              </div>
-              {backtestOptions ? (
-                <div className="backtest-notes">
-                  <div className="note-item">
-                    tick: {String(backtestOptions.availability?.tick ?? "unknown")} · dir: {String(backtestOptions.dataDirectories?.tick ?? "--")}
-                  </div>
-                  <div className="note-item">
-                    1min: {String(backtestOptions.availability?.["1min"] ?? "unknown")} · dir: {String(backtestOptions.dataDirectories?.["1min"] ?? "--")}
-                  </div>
-                  <div className="note-item">
-                    selected source: {backtestForm.executionDataSource} · {selectedExecutionDatasets.length} dataset file(s)
-                  </div>
-                  <div className="note-item">
-                    symbols: {selectedExecutionSymbols.length > 0 ? selectedExecutionSymbols.join(", ") : "--"}
-                  </div>
-                  <div className="note-item">
-                    required columns: {selectedExecutionSchema?.requiredColumns?.join(", ") ?? "--"}
-                  </div>
-                  <div className="note-item">
-                    file examples: {selectedExecutionSchema?.filenameExamples?.join(", ") ?? "--"}
-                  </div>
-                  {!selectedSymbolAvailable ? (
-                    <div className="note-item">
-                      symbol {backtestForm.symbol.trim().toUpperCase()} is not available for {backtestForm.executionDataSource}
-                    </div>
-                  ) : null}
-                  {selectedExecutionDatasets.slice(0, 3).map((dataset) => (
-                    <div key={dataset.path} className="note-item">
-                      {dataset.name} · {dataset.symbol}
-                      {dataset.format ? ` · ${dataset.format}` : ""}
-                      {dataset.fileCount ? ` · files ${dataset.fileCount}` : ""}
-                    </div>
-                  ))}
-                  {(backtestOptions.notes ?? []).map((note) => (
-                    <div key={note} className="note-item">
-                      {note}
-                    </div>
-                  ))}
-                </div>
-              ) : null}
-            </div>
-            <div className="backtest-list">
-              {backtestItems.length > 0 ? (
-                <div className="table-wrap">
-                  <table>
-                    <thead>
-                      <tr>
-                        <th>Time</th>
-                        <th>Mode</th>
-                        <th>Symbol</th>
-                        <th>Status</th>
-                        <th>Return</th>
-                        <th>DD</th>
-                      </tr>
-                    </thead>
-                    <tbody>
-                      {backtestItems.map((item) => (
-                        <tr
-                          key={item.id}
-                          className={item.id === selectedBacktest?.id ? "table-row-active" : ""}
-                          onClick={() => setSelectedBacktestId(item.id)}
-                        >
-                          <td>{formatTime(item.createdAt)}</td>
-                          <td>{String(item.parameters?.backtestMode ?? "--")}</td>
-                          <td>{String(item.parameters?.symbol ?? "--")}</td>
-                          <td>{item.status}</td>
-                          <td>{formatPercent(item.resultSummary?.return)}</td>
-                          <td>{formatPercent(item.resultSummary?.maxDrawdown)}</td>
-                        </tr>
-                      ))}
-                    </tbody>
-                  </table>
-                </div>
-              ) : (
-                <div className="empty-state">No backtests yet</div>
-              )}
-              <div className="backtest-detail-card">
-                <div className="panel-header">
-                  <div>
-                    <p className="panel-kicker">Strategy Replay</p>
-                    <h3>选中回测详情</h3>
-                  </div>
-                  <div className="range-box range-box-wrap">
-                    <span>{selectedBacktest?.status ?? "NO RUN"}</span>
-                    <span>{String(selectedBacktest?.parameters?.backtestMode ?? "--")}</span>
-                    <button
-                      type="button"
-                      className="filter-chip"
-                      disabled={!selectedBacktest?.parameters?.from || !selectedBacktest?.parameters?.to}
-                      onClick={() => {
-                        const from = Date.parse(String(selectedBacktest?.parameters?.from ?? ""));
-                        const to = Date.parse(String(selectedBacktest?.parameters?.to ?? ""));
-                        if (!Number.isFinite(from) || !Number.isFinite(to)) {
-                          return;
-                        }
-                        setChartOverrideRange({
-                          from: Math.floor(from / 1000),
-                          to: Math.floor(to / 1000),
-                          label: "Backtest Window",
-                        });
-                        setFocusNonce((value) => value + 1);
-                      }}
-                    >
-                      Use Backtest Window
-                    </button>
-                    <button
-                      type="button"
-                      className="filter-chip"
-                      disabled={!chartOverrideRange}
-                      onClick={() => setChartOverrideRange(null)}
-                    >
-                      Back To Live Window
-                    </button>
-                    <a
-                      className={`filter-chip ${selectedBacktest ? "" : "filter-chip-disabled"}`}
-                      href={selectedBacktest ? `${API_BASE}/api/v1/backtests/${selectedBacktest.id}/execution-trades.csv` : undefined}
-                    >
-                      Export Trades CSV
-                    </a>
-                  </div>
-                </div>
-                {selectedBacktest ? (
-                  <>
-                    <div className="detail-grid">
-                      <div className="detail-item">
-                        <span>Execution Source</span>
-                        <strong>{latestExecutionSource || "--"}</strong>
-                      </div>
-                      <div className="detail-item">
-                        <span>Matched Files</span>
-                        <strong>{String(latestBacktestSummary.matchedArchiveFiles ?? "--")}</strong>
-                      </div>
-                      <div className="detail-item">
-                        <span>{previewCountLabel}</span>
-                        <strong>{String(latestBacktestSummary.streamPreviewTicks ?? "--")}</strong>
-                      </div>
-                      <div className="detail-item">
-                        <span>{processedCountLabel}</span>
-                        <strong>{processedCountValue}</strong>
-                      </div>
-                      <div className="detail-item">
-                        <span>Trade Count</span>
-                        <strong>{String(latestBacktestSummary.executionTradeCount ?? "--")}</strong>
-                      </div>
-                      <div className="detail-item">
-                        <span>Closed Trades</span>
-                        <strong>{String(latestBacktestSummary.executionClosedCount ?? "--")}</strong>
-                      </div>
-                      <div className="detail-item">
-                        <span>Win Rate</span>
-                        <strong>{formatPercent(latestBacktestSummary.executionWinRate)}</strong>
-                      </div>
-                      <div className="detail-item">
-                        <span>Total PnL</span>
-                        <strong>{formatSigned(getNumber(latestBacktestSummary.executionRealizedPnL))}</strong>
-                      </div>
-                    </div>
-
-                    <div className="backtest-breakdown">
-                      <h4>Execution Trades</h4>
-                      {latestExecutionTrades.length > 0 ? (
-                        <SimpleTable
-                          columns={["Status", "Source", "Side", "Qty", "Entry", "Exit", "Exit Type", "PnL"]}
-                          rows={latestExecutionTrades.map((trade) => [
-                            String(trade.status ?? "--"),
-                            String(trade.source ?? "--"),
-                            String(trade.side ?? "--"),
-                            formatMaybeNumber(trade.quantity),
-                            `${formatMaybeNumber(trade.entryPrice)} @ ${formatTime(String(trade.entryTime ?? ""))}`,
-                            `${formatMaybeNumber(trade.exitPrice)} @ ${formatTime(String(trade.exitTime ?? ""))}`,
-                            String(trade.exitType ?? "--"),
-                            formatSigned(getNumber(trade.realizedPnL)),
-                          ])}
-                          emptyMessage="No execution trades"
-                        />
-                      ) : (
-                        <div className="empty-state empty-state-compact">No execution trades yet</div>
-                      )}
-                    </div>
-
-                    {Boolean(latestBacktestSummary.replayLedgerTrades) ? (
-                      <>
-                        <div className="backtest-breakdown">
-                          <h4>Optional Ledger Audit</h4>
-                          {Object.keys(latestReplayByReason).length > 0 ? (
-                            <SimpleTable
-                              columns={["Reason", "Trades", "Completed", "Skipped", "Entry", "Exit"]}
-                              rows={Object.entries(latestReplayByReason).map(([reason, stats]) => [
-                                reason,
-                                String(stats.trades ?? 0),
-                                String(stats.completed ?? 0),
-                                String(stats.skipped ?? 0),
-                                String(stats.skippedEntry ?? 0),
-                                String(stats.skippedExit ?? 0),
-                              ])}
-                              emptyMessage="No grouped replay stats"
-                            />
-                          ) : (
-                            <div className="empty-state empty-state-compact">No optional ledger audit data</div>
-                          )}
-                        </div>
-
-                        <div className="backtest-samples-grid">
-                          <div className="backtest-sample-panel">
-                            <h4>Completed Samples</h4>
-                            {latestReplayCompletedSamples.length > 0 ? (
-                              latestReplayCompletedSamples.map((sample, index) => (
-                                <SampleCard
-                                  key={`completed-${index}`}
-                                  sample={sample}
-                                  selected={selectedSample?.key === buildSampleKey("completed", index, sample)}
-                                  onSelect={() => {
-                                    const range = buildSampleRange(sample);
-                                    if (!range) {
-                                      return;
-                                    }
-                                    setSelectedSample({ key: buildSampleKey("completed", index, sample), sample });
-                                    setChartOverrideRange(range);
-                                    setSourceFilter("backtest");
-                                    setEventFilter("all");
-                                    setFocusNonce((value) => value + 1);
-                                  }}
-                                />
-                              ))
-                            ) : (
-                              <div className="empty-state empty-state-compact">No completed samples</div>
-                            )}
-                          </div>
-                          <div className="backtest-sample-panel">
-                            <h4>Skipped Samples</h4>
-                            {latestReplaySkippedSamples.length > 0 ? (
-                              latestReplaySkippedSamples.map((sample, index) => (
-                                <SampleCard
-                                  key={`skipped-${index}`}
-                                  sample={sample}
-                                  selected={selectedSample?.key === buildSampleKey("skipped", index, sample)}
-                                  onSelect={() => {
-                                    const range = buildSampleRange(sample);
-                                    if (!range) {
-                                      return;
-                                    }
-                                    setSelectedSample({ key: buildSampleKey("skipped", index, sample), sample });
-                                    setChartOverrideRange(range);
-                                    setSourceFilter("backtest");
-                                    setEventFilter("all");
-                                    setFocusNonce((value) => value + 1);
-                                  }}
-                                />
-                              ))
-                            ) : (
-                              <div className="empty-state empty-state-compact">No skipped samples</div>
-                            )}
-                          </div>
-                        </div>
-                      </>
-                    ) : null}
-                  </>
-                ) : (
-                  <div className="empty-state empty-state-compact">No backtest detail yet</div>
-                )}
-              </div>
-            </div>
-          </div>
-        </section>
-          </div>
+          <StrategySidePanel createBacktestRun={createBacktestRun} />
         ) : sidebarTab === 'account' ? (
-          <div className="p-4 space-y-6">
-            <section id="signals" className="panel panel-session">
-          <div className="panel-header">
-            <div>
-              <p className="panel-kicker">Signal Runtime</p>
-              <h3>信号源绑定与市场数据运行时</h3>
-            </div>
-            <div className="range-box">
-              <span>{signalCatalog?.sources?.length ?? 0} sources</span>
-              <span>{signalRuntimeSessions.length} sessions</span>
-            </div>
-          </div>
-
-          <div className="live-grid">
-            <div className="backtest-form session-form">
-              <h4>Bind Account Signal Source</h4>
-              <div className="form-grid">
-                <label className="form-field">
-                  <span>Account</span>
-                  <select value={accountSignalForm.accountId} onChange={(event) => setAccountSignalForm((current: any) => ({ ...current, accountId: event.target.value }))}>
-                    {liveAccounts.map((item) => (
-                      <option key={item.id} value={item.id}>
-                        {item.name} ({item.mode})
-                      </option>
-                    ))}
-                  </select>
-                </label>
-                <label className="form-field">
-                  <span>Source</span>
-                  <select value={accountSignalForm.sourceKey} onChange={(event) => setAccountSignalForm((current: any) => ({ ...current, sourceKey: event.target.value }))}>
-                    {(signalCatalog?.sources ?? []).map((source) => (
-                      <option key={source.key} value={source.key}>
-                        {source.name}
-                      </option>
-                    ))}
-                  </select>
-                </label>
-                <label className="form-field">
-                  <span>Role</span>
-                  <select value={accountSignalForm.role} onChange={(event) => setAccountSignalForm((current: any) => ({ ...current, role: event.target.value }))}>
-                    <option value="signal">signal</option>
-                    <option value="trigger">trigger</option>
-                    <option value="feature">feature</option>
-                  </select>
-                </label>
-                <label className="form-field">
-                  <span>Timeframe</span>
-                  <select value={accountSignalForm.timeframe} onChange={(event) => setAccountSignalForm((current: any) => ({ ...current, timeframe: event.target.value }))}>
-                    <option value="4h">4h</option>
-                    <option value="1d">1d</option>
-                  </select>
-                </label>
-                <label className="form-field">
-                  <span>Symbol</span>
-                  <input value={accountSignalForm.symbol} onChange={(event) => setAccountSignalForm((current: any) => ({ ...current, symbol: event.target.value.toUpperCase() }))} />
-                </label>
-              </div>
-              <div className="backtest-actions">
-                <ActionButton label={signalBindingAction === "account" ? "Binding..." : "Bind Account Source"} disabled={signalBindingAction !== null || !accountSignalForm.accountId} onClick={bindAccountSignalSource} />
-              </div>
-            </div>
-
-            <div className="backtest-form session-form">
-              <h4>Bind Strategy Signal Source</h4>
-              <div className="form-grid">
-                <label className="form-field">
-                  <span>Strategy</span>
-                  <select value={strategySignalForm.strategyId} onChange={(event) => setStrategySignalForm((current: any) => ({ ...current, strategyId: event.target.value }))}>
-                    {strategyOptions.map((strategy) => (
-                      <option key={strategy.value} value={strategy.value}>
-                        {strategy.label}
-                      </option>
-                    ))}
-                  </select>
-                </label>
-                <label className="form-field">
-                  <span>Source</span>
-                  <select value={strategySignalForm.sourceKey} onChange={(event) => setStrategySignalForm((current: any) => ({ ...current, sourceKey: event.target.value }))}>
-                    {(signalCatalog?.sources ?? []).map((source) => (
-                      <option key={source.key} value={source.key}>
-                        {source.name}
-                      </option>
-                    ))}
-                  </select>
-                </label>
-                <label className="form-field">
-                  <span>Role</span>
-                  <select value={strategySignalForm.role} onChange={(event) => setStrategySignalForm((current: any) => ({ ...current, role: event.target.value }))}>
-                    <option value="signal">signal</option>
-                    <option value="trigger">trigger</option>
-                    <option value="feature">feature</option>
-                  </select>
-                </label>
-                <label className="form-field">
-                  <span>Timeframe</span>
-                  <select value={strategySignalForm.timeframe} onChange={(event) => setStrategySignalForm((current: any) => ({ ...current, timeframe: event.target.value }))}>
-                    <option value="4h">4h</option>
-                    <option value="1d">1d</option>
-                  </select>
-                </label>
-                <label className="form-field">
-                  <span>Symbol</span>
-                  <input value={strategySignalForm.symbol} onChange={(event) => setStrategySignalForm((current: any) => ({ ...current, symbol: event.target.value.toUpperCase() }))} />
-                </label>
-              </div>
-              <div className="backtest-actions">
-                <ActionButton label={signalBindingAction === "strategy" ? "Binding..." : "Bind Strategy Source"} disabled={signalBindingAction !== null || !strategySignalForm.strategyId} onClick={bindStrategySignalSource} />
-              </div>
-            </div>
-          </div>
-
-          <div className="live-grid">
-            <div className="backtest-list">
-              <h4>Signal Source Catalog</h4>
-              {signalCatalog?.sources?.length ? (
-                <SimpleTable
-                  columns={["Source", "Exchange", "Type", "Roles", "Env", "Transport"]}
-                  rows={signalCatalog.sources.map((source) => [
-                    source.name,
-                    source.exchange,
-                    source.streamType,
-                    source.roles.join(", "),
-                    source.environments.join(", "),
-                    source.transport,
-                  ])}
-                  emptyMessage="No signal sources"
-                />
-              ) : (
-                <div className="empty-state empty-state-compact">No signal source catalog</div>
-              )}
-              <div className="backtest-notes">
-                {(signalCatalog?.notes ?? []).map((note) => (
-                  <div key={note} className="note-item">
-                    {note}
-                  </div>
-                ))}
-                {(signalSourceTypes ?? []).map((item) => (
-                  <div key={item.streamType} className="note-item">
-                    {item.streamType}: {item.description}
-                  </div>
-                ))}
-              </div>
-            </div>
-
-            <div className="backtest-list">
-              <h4>Current Bindings</h4>
-              <div className="backtest-breakdown">
-                <h5>Account</h5>
-                <SimpleTable
-                  columns={["Source", "Role", "Symbol", "Exchange", "Status"]}
-                  rows={accountSignalBindings.map((item) => [item.sourceName, item.role, item.symbol || "--", item.exchange, item.status])}
-                  emptyMessage="No account bindings"
-                />
-              </div>
-              <div className="backtest-breakdown">
-                <h5>Strategy</h5>
-                <SimpleTable
-                  columns={["Source", "Role", "Symbol", "Exchange", "Status"]}
-                  rows={strategySignalBindings.map((item) => [item.sourceName, item.role, item.symbol || "--", item.exchange, item.status])}
-                  emptyMessage="No strategy bindings"
-                />
-              </div>
-            </div>
-          </div>
-
-          <div className="live-grid">
-            <div className="backtest-form session-form">
-              <h4>Runtime Policy</h4>
-              <div className="form-grid">
-                <label className="form-field">
-                  <span>Trade Tick Freshness (s)</span>
-                  <input
-                    value={runtimePolicyForm.tradeTickFreshnessSeconds}
-                    onChange={(event) =>
-                      setRuntimePolicyForm((current: any) => ({ ...current, tradeTickFreshnessSeconds: event.target.value }))
-                    }
-                  />
-                </label>
-                <label className="form-field">
-                  <span>Order Book Freshness (s)</span>
-                  <input
-                    value={runtimePolicyForm.orderBookFreshnessSeconds}
-                    onChange={(event) =>
-                      setRuntimePolicyForm((current: any) => ({ ...current, orderBookFreshnessSeconds: event.target.value }))
-                    }
-                  />
-                </label>
-                <label className="form-field">
-                  <span>Signal Bar Freshness (s)</span>
-                  <input
-                    value={runtimePolicyForm.signalBarFreshnessSeconds}
-                    onChange={(event) =>
-                      setRuntimePolicyForm((current: any) => ({ ...current, signalBarFreshnessSeconds: event.target.value }))
-                    }
-                  />
-                </label>
-                <label className="form-field">
-                  <span>Runtime Quiet (s)</span>
-                  <input
-                    value={runtimePolicyForm.runtimeQuietSeconds}
-                    onChange={(event) =>
-                      setRuntimePolicyForm((current: any) => ({ ...current, runtimeQuietSeconds: event.target.value }))
-                    }
-                  />
-                </label>
-                <label className="form-field">
-                  <span>Paper Start Timeout (s)</span>
-                  <input
-                    value={runtimePolicyForm.paperStartReadinessTimeoutSeconds}
-                    onChange={(event) =>
-                      setRuntimePolicyForm((current: any) => ({
-                        ...current,
-                        paperStartReadinessTimeoutSeconds: event.target.value,
-                      }))
-                    }
-                  />
-                </label>
-              </div>
-              <div className="backtest-actions">
-                <ActionButton
-                  label={runtimePolicyAction ? "Saving..." : "Save Runtime Policy"}
-                  disabled={runtimePolicyAction}
-                  onClick={updateRuntimePolicy}
-                />
-              </div>
-              <div className="backtest-notes">
-                <div className="note-item">
-                  active policy: tick {runtimePolicy?.tradeTickFreshnessSeconds ?? "--"}s · book {runtimePolicy?.orderBookFreshnessSeconds ?? "--"}s ·
-                  bar {runtimePolicy?.signalBarFreshnessSeconds ?? "--"}s
-                </div>
-                <div className="note-item">
-                  quiet {runtimePolicy?.runtimeQuietSeconds ?? "--"}s · paper preflight {runtimePolicy?.paperStartReadinessTimeoutSeconds ?? "--"}s
-                </div>
-              </div>
-            </div>
-
-            <div className="backtest-form session-form">
-              <h4>Create Runtime Session</h4>
-              <div className="form-grid">
-                <label className="form-field">
-                  <span>Account</span>
-                  <select value={signalRuntimeForm.accountId} onChange={(event) => setSignalRuntimeForm((current: any) => ({ ...current, accountId: event.target.value }))}>
-                    {liveAccounts.map((item) => (
-                      <option key={item.id} value={item.id}>
-                        {item.name} ({item.mode})
-                      </option>
-                    ))}
-                  </select>
-                </label>
-                <label className="form-field">
-                  <span>Strategy</span>
-                  <select value={signalRuntimeForm.strategyId} onChange={(event) => setSignalRuntimeForm((current: any) => ({ ...current, strategyId: event.target.value }))}>
-                    {strategyOptions.map((strategy) => (
-                      <option key={strategy.value} value={strategy.value}>
-                        {strategy.label}
-                      </option>
-                    ))}
-                  </select>
-                </label>
-              </div>
-              <div className="backtest-actions">
-                <ActionButton label={signalRuntimeAction === "create" ? "Creating..." : "Create Runtime Session"} disabled={signalRuntimeAction !== null || !signalRuntimeForm.accountId || !signalRuntimeForm.strategyId} onClick={createSignalRuntimeSession} />
-              </div>
-              <div className="detail-grid">
-                <div className="detail-item">
-                  <span>Plan Ready</span>
-                  <strong>{boolLabel(signalRuntimePlan?.ready)}</strong>
-                </div>
-                <div className="detail-item">
-                  <span>Required</span>
-                  <strong>{String((signalRuntimePlan?.requiredBindings as unknown[] | undefined)?.length ?? 0)}</strong>
-                </div>
-                <div className="detail-item">
-                  <span>Matched</span>
-                  <strong>{String((signalRuntimePlan?.matchedBindings as unknown[] | undefined)?.length ?? 0)}</strong>
-                </div>
-                <div className="detail-item">
-                  <span>Missing</span>
-                  <strong>{String((signalRuntimePlan?.missingBindings as unknown[] | undefined)?.length ?? 0)}</strong>
-                </div>
-              </div>
-              <div className="backtest-notes">
-                <div className="note-item">runtime adapters: {signalRuntimeAdapters.map((item) => item.key).join(", ") || "--"}</div>
-                {((signalRuntimePlan?.missingBindings as unknown[] | undefined) ?? []).slice(0, 4).map((item, index) => (
-                  <div key={index} className="note-item">
-                    missing: {JSON.stringify(item)}
-                  </div>
-                ))}
-              </div>
-            </div>
-
-            <div className="backtest-list">
-              <h4>Runtime Sessions</h4>
-              {signalRuntimeSessions.length > 0 ? (
-                <>
-                  <div className="table-wrap">
-                    <table>
-                      <thead>
-                        <tr>
-                          <th>Session</th>
-                          <th>Status</th>
-                          <th>Adapter</th>
-                          <th>Subs</th>
-                          <th>Heartbeat</th>
-                          <th>Action</th>
-                        </tr>
-                      </thead>
-                      <tbody>
-                        {signalRuntimeSessions.map((session) => (
-                          <tr
-                            key={session.id}
-                            className={session.id === selectedSignalRuntime?.id ? "table-row-active" : ""}
-                            onClick={() => setSelectedSignalRuntimeId(session.id)}
-                          >
-                            <td>{shrink(session.id)}</td>
-                            <td>{session.status}</td>
-                            <td>{session.runtimeAdapter || "--"}</td>
-                            <td>{String(session.subscriptionCount)}</td>
-                            <td>{formatTime(String(session.state?.lastHeartbeatAt ?? ""))}</td>
-                            <td>
-                              <div className="inline-actions">
-                                <ActionButton
-                                  label={signalRuntimeAction === `${session.id}:start` ? "Starting..." : "Start"}
-                                  disabled={signalRuntimeAction !== null || session.status === "RUNNING"}
-                                  onClick={() => runSignalRuntimeAction(session.id, "start")}
-                                />
-                                <ActionButton
-                                  label={signalRuntimeAction === `${session.id}:stop` ? "Stopping..." : "Stop"}
-                                  variant="ghost"
-                                  disabled={signalRuntimeAction !== null || session.status === "STOPPED"}
-                                  onClick={() => runSignalRuntimeAction(session.id, "stop")}
-                                />
-                              </div>
-                            </td>
-                          </tr>
-                        ))}
-                      </tbody>
-                    </table>
-                  </div>
-
-                  <div className="backtest-detail-card">
-                    <div className="panel-header">
-                      <div>
-                        <p className="panel-kicker">Signal Session</p>
-                        <h3>选中 Runtime Session 详情</h3>
-                      </div>
-                      <div className="range-box">
-                        <span>{selectedSignalRuntime?.status ?? "NO SESSION"}</span>
-                        <span>{selectedSignalRuntime?.runtimeAdapter ?? "--"}</span>
-                      </div>
-                    </div>
-                    {selectedSignalRuntime ? (
-                      <>
-                        <div className="detail-grid">
-                          <div className="detail-item">
-                            <span>Session ID</span>
-                            <strong>{shrink(selectedSignalRuntime.id)}</strong>
-                          </div>
-                          <div className="detail-item">
-                            <span>Account</span>
-                            <strong>{shrink(selectedSignalRuntime.accountId)}</strong>
-                          </div>
-                          <div className="detail-item">
-                            <span>Strategy</span>
-                            <strong>{shrink(selectedSignalRuntime.strategyId)}</strong>
-                          </div>
-                          <div className="detail-item">
-                            <span>Transport</span>
-                            <strong>{selectedSignalRuntime.transport || "--"}</strong>
-                          </div>
-                          <div className="detail-item">
-                            <span>Health</span>
-                            <strong>{String(selectedSignalRuntimeState.health ?? "--")}</strong>
-                          </div>
-                          <div className="detail-item">
-                            <span>Signal Events</span>
-                            <strong>{String(Math.trunc(getNumber(selectedSignalRuntimeState.signalEventCount) ?? 0))}</strong>
-                          </div>
-                          <div className="detail-item">
-                            <span>Heartbeat</span>
-                            <strong>{formatTime(String(selectedSignalRuntimeState.lastHeartbeatAt ?? ""))}</strong>
-                          </div>
-                          <div className="detail-item">
-                            <span>Last Event</span>
-                            <strong>{formatTime(String(selectedSignalRuntimeState.lastEventAt ?? ""))}</strong>
-                          </div>
-                          <div className="detail-item">
-                            <span>Source States</span>
-                            <strong>{String(Object.keys(selectedSignalRuntimeSourceStates).length)}</strong>
-                          </div>
-                          <div className="detail-item">
-                            <span>Plan Ready</span>
-                            <strong>{boolLabel(selectedSignalRuntimePlan.ready)}</strong>
-                          </div>
-                        </div>
-
-                        <div className="backtest-breakdown">
-                          <h4>Subscriptions</h4>
-                          <SimpleTable
-                            columns={["Source", "Role", "Symbol", "Channel", "Adapter"]}
-                            rows={selectedSignalRuntimeSubscriptions.map((item) => [
-                              String(item.sourceKey ?? "--"),
-                              String(item.role ?? "--"),
-                              String(item.symbol ?? "--"),
-                              String(item.channel ?? "--"),
-                              String(item.adapterKey ?? "--"),
-                            ])}
-                            emptyMessage="No subscriptions"
-                          />
-                        </div>
-
-                        <div className="backtest-breakdown">
-                          <h4>Signal Bars</h4>
-                          {selectedSignalRuntimeSignalBars.length > 0 ? (
-                            <div className="chart-shell">
-                              <SignalBarChart candles={selectedSignalRuntimeSignalBars} />
-                            </div>
-                          ) : (
-                            <div className="empty-state empty-state-compact">No 4h/1d signal bars cached yet</div>
-                          )}
-                        </div>
-
-                        <div className="backtest-breakdown">
-                          <h4>Signal States</h4>
-                          <div className="backtest-notes">
-                            {Object.entries(selectedSignalBarStates).length > 0 ? (
-                              Object.entries(selectedSignalBarStates).map(([key, value]) => {
-                                const state = getRecord(value);
-                                const current = getRecord(state.current);
-                                const prevBar1 = getRecord(state.prevBar1);
-                                const prevBar2 = getRecord(state.prevBar2);
-                                return (
-                                  <div key={key} className="note-item">
-                                    {[
-                                      key,
-                                      `tf=${String(state.timeframe ?? "--")}`,
-                                      `bars=${String(state.barCount ?? "--")}`,
-                                      `ma20=${formatMaybeNumber(state.ma20)}`,
-                                      `atr14=${formatMaybeNumber(state.atr14)}`,
-                                      `t-1=${formatMaybeNumber(prevBar1.open)}/${formatMaybeNumber(prevBar1.high)}/${formatMaybeNumber(prevBar1.low)}/${formatMaybeNumber(prevBar1.close)}`,
-                                      `t-2=${formatMaybeNumber(prevBar2.open)}/${formatMaybeNumber(prevBar2.high)}/${formatMaybeNumber(prevBar2.low)}/${formatMaybeNumber(prevBar2.close)}`,
-                                      `current=${formatMaybeNumber(current.open)}/${formatMaybeNumber(current.high)}/${formatMaybeNumber(current.low)}/${formatMaybeNumber(current.close)}`,
-                                    ].join(" · ")}
-                                  </div>
-                                );
-                              })
-                            ) : (
-                              <div className="empty-state empty-state-compact">No signal states yet</div>
-                            )}
-                          </div>
-                        </div>
-
-                        <div className="backtest-breakdown">
-                          <h4>Runtime Timeline</h4>
-                          <div className="backtest-notes">
-                            {buildTimelineNotes(selectedSignalRuntimeTimeline).map((line) => (
-                              <div key={line} className="note-item">
-                                {line}
-                              </div>
-                            ))}
-                          </div>
-                        </div>
-
-                        <div className="backtest-breakdown">
-                          <h4>Last Event Summary</h4>
-                          <div className="backtest-notes">
-                            {Object.entries(selectedSignalRuntimeLastSummary).length > 0 ? (
-                              Object.entries(selectedSignalRuntimeLastSummary).map(([key, value]) => (
-                                <div key={key} className="note-item">
-                                  {key}: {typeof value === "object" ? JSON.stringify(value) : String(value)}
-                                </div>
-                              ))
-                            ) : (
-                              <div className="empty-state empty-state-compact">No event summary yet</div>
-                            )}
-                          </div>
-                        </div>
-
-                        <div className="backtest-breakdown">
-                          <h4>Source States</h4>
-                          <div className="backtest-notes">
-                            {Object.entries(selectedSignalRuntimeSourceStates).length > 0 ? (
-                              Object.entries(selectedSignalRuntimeSourceStates).slice(0, 8).map(([key, value]) => (
-                                <div key={key} className="note-item">
-                                  {key}: {typeof value === "object" ? JSON.stringify(value) : String(value)}
-                                </div>
-                              ))
-                            ) : (
-                              <div className="empty-state empty-state-compact">No source states yet</div>
-                            )}
-                          </div>
-                        </div>
-                      </>
-                    ) : (
-                      <div className="empty-state empty-state-compact">No runtime session selected</div>
-                    )}
-                  </div>
-                </>
-              ) : (
-                <div className="empty-state empty-state-compact">No runtime sessions</div>
-              )}
-            </div>
-          </div>
-        </section>
-          </div>
+          <AccountSidePanel
+            bindAccountSignalSource={bindAccountSignalSource}
+            bindStrategySignalSource={bindStrategySignalSource}
+            updateRuntimePolicy={updateRuntimePolicy}
+            createSignalRuntimeSession={createSignalRuntimeSession}
+            runSignalRuntimeAction={runSignalRuntimeAction}
+          />
         ) : null
       }
       dockContent={
